09-04-2025  appointment files and feedback changes

+++ springapp/src/main/java/com/examly/springapp/service/FeedbackServiceImpl.java
@@ -52,7 +52,7 @@ public class FeedbackServiceImpl implements FeedbackService {
     
 
     @Override
-    public List<Feedback> getFeedbackByUserId(Long userId) {
+    public List<Feedback> getFeedbackByUserId(int userId) {
         List<Feedback> feedbackList=feedbackRepo.findByUserId(userId);
         return feedbackList;
     }
